remove `_jst.Pop`
