+++ python/paddle/base/layers/math_op_patch.py
@@ -67,6 +67,24 @@ EXPRESSION_MAP = {
 _already_patch_variable = False
 
 
+# TODO(liym27): A better way to slice tensor array.
+#  Maybe support start == end for slice op.
+def _slice_tensor_array(array, start, end):
+    from paddle.static.nn import cond
+    from paddle.tensor import create_array
+
+    def true_fn():
+        null_array = create_array("float32")
+        return null_array
+
+    def false_fn(array, start, end):
+        new_array = array[start:end]
+        return new_array
+
+    new_array = cond(start == end, true_fn, lambda: false_fn(array, start, end))
+    return new_array
+
+
 def monkey_patch_variable():
     def unique_tmp_name():
         return default_main_program()._name_generator.generate("tmp")
@@ -360,7 +378,9 @@ def monkey_patch_variable():
         Returns:
             Variable: self[index]
         """
-        from paddle.jit.dy2static.convert_operators import _run_paddle_pop
+        import paddle
+        from paddle.static.nn import while_loop
+        from paddle.tensor import fill_constant
 
         if self.type != core.VarDesc.VarType.LOD_TENSOR_ARRAY:
             raise TypeError(
@@ -368,7 +388,41 @@ def monkey_patch_variable():
                     self.type
                 )
             )
-        return _run_paddle_pop(self, *args)
+        if len(args) == 0:
+            idx = -1
+        else:
+            idx = args[0]
+
+        assert isinstance(idx, int)
+
+        def cond(i, new_array):
+            return paddle.less_than(i, arr_len)
+
+        def body(i, new_array):
+            item = paddle.tensor.array_read(array=self, i=i)
+            paddle.tensor.array_write(
+                item, paddle.tensor.array_length(new_array), new_array
+            )
+
+            i = paddle.increment(i)
+            return i, new_array
+
+        arr_len = paddle.tensor.array_length(self)
+        if idx < 0:
+            idx = idx + arr_len
+        else:
+            idx = fill_constant(shape=[1], dtype="int64", value=idx)
+
+        pop_item = paddle.tensor.array_read(self, idx)
+
+        tmp = paddle.assign(self)
+        new_array = _slice_tensor_array(tmp, 0, idx)
+        i = idx + 1
+
+        _, new_array = while_loop(cond, body, [i, new_array])
+        paddle.assign(new_array, output=self)
+
+        return pop_item
 
     def _scalar_op_(var, scale, bias):
         block = current_block(var)
