+++ python/paddle/jit/dy2static/convert_operators.py
@@ -754,28 +754,6 @@ def convert_print(*objects, sep=' ', end='\n', file=None, flush=False):
     print(*objects, sep=sep, end=end, file=file, flush=flush)
 
 
-def convert_pop(target, *args):
-    """
-    A function representation of a Python pop statement for a list or dict.
-
-    Args:
-        target(list|dict|Tensor): A variable to pop item from.
-        *args(tuple): index or default value to parse.
-
-    Returns:
-        A item poped from target.
-    """
-
-    is_variable = isinstance(target, (Variable, Value))
-    if is_variable:
-        is_tensor_array = target.type == core.VarDesc.VarType.LOD_TENSOR_ARRAY
-
-    if is_variable and is_tensor_array:
-        return _run_paddle_pop(target, *args)
-    else:
-        return _run_python_pop(target, *args)
-
-
 @contextmanager
 def convert_auto_cast(
     enable=True,
@@ -807,73 +785,3 @@ def convert_auto_cast(
     amp_records[current_block_idx].append(
         (amp_options, start_op_idx, end_op_idx)
     )
-
-
-def _run_paddle_pop(array, *args):
-    if len(args) == 0:
-        idx = -1
-    else:
-        idx = args[0]
-
-    assert isinstance(idx, int)
-
-    def cond(i, new_array):
-        return paddle.less_than(i, arr_len)
-
-    def body(i, new_array):
-        item = paddle.tensor.array_read(array=array, i=i)
-        paddle.tensor.array_write(
-            item, paddle.tensor.array_length(new_array), new_array
-        )
-
-        i = paddle.increment(i)
-        return i, new_array
-
-    arr_len = paddle.tensor.array_length(array)
-    if idx < 0:
-        idx = idx + arr_len
-    else:
-        from paddle.tensor import fill_constant
-
-        idx = fill_constant(shape=[1], dtype="int64", value=idx)
-
-    pop_item = paddle.tensor.array_read(array, idx)
-
-    tmp = paddle.assign(array)
-    new_array = _slice_tensor_array(tmp, 0, idx)
-    i = idx + 1
-    from paddle.static.nn import while_loop
-
-    _, new_array = while_loop(cond, body, [i, new_array])
-    paddle.assign(new_array, output=array)
-
-    return pop_item
-
-
-# TODO(liym27): A better way to slice tensor array.
-#  Maybe support start == end for slice op.
-def _slice_tensor_array(array, start, end):
-    def true_fn():
-        null_array = paddle.tensor.create_array("float32")
-        return null_array
-
-    def false_fn(array, start, end):
-        new_array = array[start:end]
-        return new_array
-
-    new_array = paddle.static.nn.cond(
-        start == end, true_fn, lambda: false_fn(array, start, end)
-    )
-    return new_array
-
-
-def _run_python_pop(target, *args):
-    # 1. pop for a dict
-    if len(args) == 2:
-        idx, default = args
-        return target.pop(idx, default)
-
-    # 2. pop for a list or dict
-    else:
-        idx = args[0] if args else -1
-        return target.pop(idx)
